Added movement for pawn. Get rid of is_path_clear
Now most movement validation will be done inside valid_move() instead of
having a function for movement and a function for path checking. This is
because a pawn's movement was too complicated as how it moves depends on
the state of the board around it (and what it can take).